LazyTypeEqualByComparer when types are not sealed.
