+++ Gu.State/EqualBy/Comparers/LazyTypeEqualByComparer.cs
@@ -1,0 +1,18 @@
+﻿namespace Gu.State
+{
+    internal class LazyTypeEqualByComparer : EqualByComparer
+    {
+        internal static readonly LazyTypeEqualByComparer Default = new LazyTypeEqualByComparer();
+
+        private LazyTypeEqualByComparer()
+        {
+        }
+
+        public override bool Equals(object x, object y, MemberSettings settings, ReferencePairCollection referencePairs)
+        {
+            return TryGetEitherNullEquals(x, y, out var result)
+                ? result
+                : settings.GetEqualByComparer(x.GetType()).Equals(x, y, settings, referencePairs);
+        }
+    }
+}
